Fix type consistency in EEPROM and related modules
- Change wlIdxNxtWr from uint32_t to uint8_t with UINT8_MAX sentinel
- Remove unused pIdx parameter from eepromWriteWL()
- Fix signed/unsigned type mismatches in eeprom.c
- Add explicit casts for i2cDataWrite address.lsb calls
- Update loop counters to appropriate unsigned types
- Add 'u' suffix to constants for type consistency

+++ src/driver_SERCOM.c
@@ -388,7 +388,7 @@ void i2cBusRecovery(Sercom *sercom, const uint8_t grp, const uint8_t sdaPin,
   i2cmCommon(sercom);
 }
 
-I2CM_Status_t i2cActivate(Sercom *sercom, const uint8_t addr) {
+I2CM_Status_t i2cActivate(Sercom *sercom, const uint32_t addr) {
   uint32_t      t = timerMicros();
   I2CM_Status_t s = I2CM_SUCCESS;
